requirements.txt
# FastAPI and web framework
fastapi==0.115.12
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Data processing
pandas==2.0.3

# Geographic data processing
pgeocode==0.4.0

# Caching and background tasks
redis==5.0.1
celery==5.3.4

# AI services (optional - add API keys to use)
# anthropic==0.7.8
# openai==1.3.5

# Payment processing (optional)
# stripe==7.8.0

# Monitoring (optional)
# sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Email support for 2FA
fastapi-mail==1.4.1
aiosmtplib>=2.0,<3.0
email-validator==2.1.0

# Template engine for email templates
Jinja2==3.1.2

# Additional email dependencies for 2FA
# Note: smtplib and email.mime are built-in Python modules, no installation needed

# Legacy Flask support (for existing webhook_server.py)
Flask==2.3.3
gunicorn==21.2.0
